Codes/bios_codes/requirements.txt
astroid==2.2.5
autopep8==1.4.4
isort==4.3.20
joblib==0.13.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.16.4
protobuf==3.8.0
pycodestyle==2.5.0
pylint==2.3.1
pytorch-ignite==0.2.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboardX==1.7
typed-ast==1.4.0
wrapt==1.11.2
